requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djangorestframework-simplejwt
channels
channels-redis
psycopg2-binary
uvicorn[standard]
